chore(runtime): move `InstanceConfig` to wasmer
