+++ dot/rpc/http_test.go
@@ -413,7 +413,7 @@ func newCoreServiceTest(t *testing.T) *core.Service {
 	err = cfg.Keystore.Acco.Insert(kp)
 	require.NoError(t, err)
 
-	var rtCfg runtime.InstanceConfig
+	var rtCfg wasmer.InstanceConfig
 
 	rtCfg.Storage = rtstorage.NewTrieState(genTrie)
