+++ lib/runtime/wasmer/config.go
@@ -1,0 +1,23 @@
+// Copyright 2022 ChainSafe Systems (ON)
+// SPDX-License-Identifier: LGPL-3.0-only
+
+package wasmer
+
+import (
+	"github.com/ChainSafe/gossamer/internal/log"
+	"github.com/ChainSafe/gossamer/lib/common"
+	"github.com/ChainSafe/gossamer/lib/keystore"
+	"github.com/ChainSafe/gossamer/lib/runtime"
+)
+
+// InstanceConfig represents a runtime instance configuration
+type InstanceConfig struct {
+	Storage     runtime.Storage
+	Keystore    *keystore.GlobalKeystore
+	LogLvl      log.Level
+	Role        common.Roles
+	NodeStorage runtime.NodeStorage
+	Network     runtime.BasicNetwork
+	Transaction runtime.TransactionState
+	CodeHash    common.Hash
+}
